+++ Changelog.md
@@ -1,3 +1,9 @@
+### 0.6.1.0
+
+- Handler *extractRequestStatusFromFullResponse* was replaced with handler
+  *extractExceptionFromFullResponse*. The new handler contains the error message
+  if the subrequest was terminated by an exception.
+
 ### 0.6.0.0
 
 - Build EDE module with *prettyprinter* automatically.